Update se4g_JupyterDash_demo -Angix.ipynb
Now the Jupyter DashBoard works! I have modified the snippet of code in such a way that the user select firstly the country, and then the available pollutants for the selected country. I have also add a barplot that shows for the selected country and pollutant the mean, the max and the average. On local it works! :)